NO-JIRA: Remove perror calls in messenger library.
No library should *ever* write unsolicited output to stderr or stdout without
being explicitly configured to do so. If the app is piping its output somewhere
it can break the application. At best it is confusing and useless, since almost
no application queries its own stdout/stderr for error status.

The connection closure in these two cases is reported to the calling app by
closing the transport. If additional error information is required it should be
set on the transport condition.