PROTON-1277: pn_connection_engine to allow external buffers
Extended the pn_connection_engine interface with use_read/write_buffer functions
to allow an external buffer to be used. API allows user-controlled buffer
re-allocation.

I tried for a "neater" interface but it didn't work out. For async write you
have to separate "get the data" and "all done with the data" so you can't really
make it simpler than it is, and you need to be able to grow both buffers so you
can't really do much to simplify the read side. For reads the transport moves
data around to accumulate complete frames, so you still have to ask "where is
the readable bit" even if you supplied the buffer.

No implementation yet. I feel like we need a clearer idea of what we are trying
to achieve before we commit. This proposal is backwards compatible which takes
some of the pressure off about deciding when to add it.