PROTON-1259: C pn_container_t interface.
Interface and implementation strategy for a C "container" based on the
connection_engine, eventually intended to replace the reactor and allow multiple
IO and threading implementations..

Stripped down to essential features to minimize the burden on implementors but
support implementing existing C++, python, Go etc. containers on top of it.
Extra convenience can be done in the the binding language where they are easier
to implement and can use native language features.